// src/types/hosts.ts
/*
 * Copyright 2022 Nightingale Team
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */
export interface IManagedHost {
  id: number;
  host_ident: string;
  ssh_ip: string;
  ssh_port: number;
  ssh_user: string;
  auth_method: 'key' | 'password';
  credential_ref: string;
  status: 'pending' | 'active' | 'failed' | 'disabled';
  note?: string;
  sudo_required: boolean;
  create_at: number;
  update_at: number;
  create_by: string;
  update_by: string;
}

export interface IManagedHostQuery {
  limit: number;
  offset: number;
  query: string;
}

export interface IManagedHostResponse {
  list: IManagedHost[];
  total: number;
}